+++ routes/web.php
@@ -13,6 +13,7 @@ use Illuminate\Support\Facades\Route;
 |
 */
 Route::get('download/file/{roomFile}/{filename?}','FileController@show')->name('download.file')->middleware('signed');
+Route::get('download/attendance/{meeting}', 'MeetingController@attendance')->middleware('auth:users,ldap');
 
 if (!env('DISABLE_CATCHALL_ROUTES')) {
     Route::any('/{any}', 'ApplicationController@index')->where('any', '.*');